// src/generated/nexus-typegen.ts
export interface NexusGenInputs {
  CardInput: { // input type
    numOfGames: number; // Int!
    sortBy: NexusGenEnums['StatSortFields']; // StatSortFields!
    take: number; // Int!
  }
}

export interface NexusGenEnums {
  Decision: "L" | "W"
  GameType: "P" | "R"
  HighlightType: "CONDENSED" | "MILESTONE" | "RECAP"
  PeriodType: "OVERTIME" | "REGULAR" | "SHOOTOUT"
  Position: "C" | "D" | "G" | "L" | "NA" | "R"
  Role: "ADMIN" | "MODERATOR" | "USER"
  RosterStatus: "I" | "N" | "Y"
  ShootsCatches: "L" | "R"
  StatSortFields: "assists" | "blocked" | "evenSavePct" | "evenSaves" | "evenTimeOnIce" | "faceOffPct" | "faceOffsTaken" | "faceOffWins" | "goalsAgainstAverage" | "giveaways" | "goals" | "goalsAgainst" | "goalsFor" | "hits" | "hitsAgainstPerGame" | "hitsForPerGame" | "losses" | "otLosses" | "penaltyKillPct" | "plusMinus" | "penaltyMinutes" | "points" | "powerPlayAssists" | "powerPlayGoals" | "powerPlayPct" | "powerPlayPoints" | "powerPlaySavePct" | "powerPlaySaves" | "powerPlaySavePct" | "powerPlayShotsAgainst" | "powerPlayShotsAgainst" | "powerPlayTimeOnIce" | "savePct" | "saves" | "savesPerGame" | "savePct" | "shots" | "shotsAgainst" | "shotsAgainstPerGame" | "shotsForPerGame" | "shutouts" | "shortHandedAssists" | "shortHandedGoals" | "shortHandedPoints" | "shortHandedSavePct" | "shortHandedSaves" | "shortHandedSavePct" | "shortHandedShotsAgainst" | "shortHandedShotsAgainst" | "shortHandedTimes" | "shortHandedTimeOnIce" | "takeaways" | "timeOnIce" | "timeOnIcePerGame" | "wins" | "winPct"
  VideoDataType: "GOAL" | "SHOT"
}